build: Add site to main repository
This makes it easier to update the cssjanus.github.io repository,
as well as making it easier to test new changes by always having a
master-based demo page ready for use.

To use the demo locally, just open the 'site/' directory in a browser.
To update the cssjanus.github.io repository, run:

 $ grunt site --dest ../cssjanus.github.io/

(Where --dest points to where your checkout of the site repo is)