test: Add test for 3-values border-radius
Should help increase the code coverage and overall stability.
This wasn't previously done because the test runner insisted
on roundtripping all test cases, which isn't possible in this case.

> input: 1px 2px 3px      -> output: 2px 1px 2px 3px
> // Same as:
> input: 1px 2px 3px 2px  -> output: 2px 1px 2px 3px

and:

> input: 2px 1px 2px 3px  -> output: 1px 2px 3px 2px
> // of which the output is effectively the same as "1px 2px 3px"

When 3 values are given in CSS, the fourth value is implied to be the same
as the second value:
> (top, [right=top], [bottom=top], [left=right])

Add a "roundtrip" setting to disable roundtripping where needed.