test: Fix invalid `end` event handler on https.get() return
There are not many cases where this is actually reachable, which is
why the existing code worked fine. For example, if we change one of
the URLs to be a 404 error, it gets handled in the `handleResponse`
callback by the status code condition.

Change this to an `error` handler instead, as per the Node.js docs [1].
This can happen for DNS/TCP errors, or internal HTTP response parsing.

Thanks @emarteca!

[1] https://nodejs.org/api/http.html#httprequesturl-options-callback

Fixes https://github.com/cssjanus/cssjanus/issues/80.