Replace WikimediaUI Base with Codex design tokens
This patch replaces the WikimediaUI Base library with the new
Codex design tokens library.
The new library is mostly a drop-in replacement for the old one.
But there are deprecated variable names (token keys) that need to be
adapted to the new ones.
See https://gerrit.wikimedia.org/r/plugins/gitiles/design/codex/+/refs/heads/main/packages/codex-design-tokens/src/deprecated-aliases-wikimedia-ui-base.json
for the full list of deprecated names with equal values.

Replacing also a number of static values mainly on properties, where
most other occurrences had already majority of variable usage. See
`font-weight` for example.

Please note that the only minor “step back” (to static values) is the
replacement of very few static background colors that are going to be
added to Codex in one of the following releases.
Also note, that `color-subtle` was agreed on by systems designers to
carry darker `#54595d` value. For now approaching it by direct
replacement as the predecessor wasn't used in prominent places. We can
work in a separate patch if a 1:1 value variable were greatly desired.
Lastly adding a small margin in label elements to make the checkboxes
easier targetable.

Bug: T331403
Bug: T334934
Change-Id: I102a641725a89a92a50f5f72c5033c8200aa4e45